Fix issues in fire-assets test
