+++ test/main/texture.js
@@ -2,7 +2,7 @@ var Fs = require('fire-fs');
 var Path = require('fire-path');
 var Url = require('fire-url');
 
-var AssetDBUtils = require('./utils');
+var AssetDBUtils = require('../utils');
 
 //
 describe('texture', function () {
@@ -26,8 +26,8 @@ describe('texture', function () {
             var uuid = Editor.assetdb.urlToUuid(url);
             var basename = Path.basename(url);
 
-            var jsonPath = Editor.assetdb._uuid2importPath(uuid);
-            var filePath = Path.join(Path.dirname(Editor.assetdb._uuid2importPath(uuid)), uuid, basename);
+            var jsonPath = Editor.assetdb._uuidToImportPathNoExt( uuid ) + '.json';
+            var filePath = Path.join(Path.dirname(jsonPath), uuid, basename);
 
             expect( Fs.existsSync( jsonPath ) )
                 .to.be.equal(true);
